tests/functional/requirements.txt
aiohttp==3.9.5
elasticsearch==8.14.0
pydantic-settings==2.3.4
pytest==8.2.2
pytest-asyncio==0.23.7
redis==5.0.7